+++ system-events/src/Configuration/ISlackApiConfiguration.cs
@@ -1,0 +1,7 @@
+namespace SystemEvents.Configuration
+{
+    public interface ISlackApiConfiguration
+    {
+        public string SlackBotUserOAuthAccessToken { get; }
+    }
+}