+++ system-events/src/Controllers/SystemEventsController.cs
@@ -4,12 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System.Threading;
 using SystemEvents.Models;
-using Nest;
-using SystemEvents.Configuration;
 using SystemEvents.Utils.Interfaces;
-using System.Collections.Generic;
-using System.Linq;
-using SystemEvents.Enums;
 
 namespace SystemEvents.Controllers
 {
@@ -17,26 +12,38 @@ namespace SystemEvents.Controllers
     public class SystemEventsController : ControllerBase
     {
         private readonly ILogger<SystemEventsController> _logger;
-        private readonly IMonitoredElasticsearchClient _esClient;
-        private readonly IElasticsearchTimeStampFactory _timeStampFactory;
-        private readonly IElasticsearchClientConfiguration _esClientConfiguration;
-        private readonly ICategorySubscriptionNotifier _categorySubscriptionNotifier;
-        private readonly IAdvanceConfiguration _advanceConfiguration;
+        private readonly ISystemEventSender _systemEventSender;
 
         public SystemEventsController(
             ILogger<SystemEventsController> logger,
-            IMonitoredElasticsearchClient esClient,
-            IElasticsearchTimeStampFactory timeStampFactory,
-            IElasticsearchClientConfiguration esClientConfiguration,
-            IAdvanceConfiguration advanceConfiguration,
-            ICategorySubscriptionNotifier categorySubscriptionNotifier = null)
+            ISystemEventSender systemEventSender)
         {
             _logger                       = logger ?? throw new ArgumentNullException(nameof(logger));
-            _esClient                     = esClient ?? throw new ArgumentNullException(nameof(esClient));
-            _timeStampFactory             = timeStampFactory ?? throw new ArgumentNullException(nameof(timeStampFactory));
-            _esClientConfiguration        = esClientConfiguration ?? throw new ArgumentNullException(nameof(esClientConfiguration));
-            _advanceConfiguration         = advanceConfiguration ?? throw new ArgumentNullException(nameof(advanceConfiguration));
-            _categorySubscriptionNotifier = categorySubscriptionNotifier;
+            _systemEventSender            = systemEventSender ?? throw new ArgumentNullException(nameof(systemEventSender));
+        }
+
+        /// <summary>
+        /// Gets a system event by id
+        /// </summary>
+        /// <returns>
+        ///   <seealso cref="Task{ActionResult{SystemEventElasticsearchDocument}}"/>
+        /// </returns>
+        [HttpGet]
+        [Route("event")]
+        public async Task<ActionResult<SystemEventElasticsearchDocument>> GetEventById(string eventId, CancellationToken cancellationToken)
+        {
+            if (string.IsNullOrWhiteSpace(eventId))
+            {
+                return BadRequest($"{nameof(eventId)} can not be null or whitespace");
+            }
+
+            var result = await _systemEventSender.Get(eventId, cancellationToken);
+            if (result == null)
+            {
+                return BadRequest($"There is no event with id {eventId}");
+            }
+
+            return Ok(result);
         }
 
         /// <summary>
@@ -49,39 +56,14 @@ namespace SystemEvents.Controllers
         [Route("event")]
         public async Task<ActionResult<EventResponse>> Create([FromBody] EventRequestModel model, CancellationToken cancellationToken)
         {
-            if (!IsValid(model, out string reason))
+            if (!_systemEventSender.Validate(model, out string reason))
             {
                 return BadRequest($"Invalid request: {reason}");
             }
 
-            var response = await CreateSystemEvent(model, cancellationToken);
-            if (!response.IsValid)
-            {
-                _logger.LogDebug(response.DebugInformation);
-                _logger.LogError(response.OriginalException, "System event was not created");
-                return BadRequest("System event was not created");
-            }
-
-            if (_categorySubscriptionNotifier == null)
-            {
-                return Ok(new EventResponse{
-                    EventId = response.Id
-                });
-            }
-
-            // Notify about this event
-            try
-            {
-                var document = await _esClient.GetAsync(response.Id, cancellationToken);
-                await _categorySubscriptionNotifier.OnEventCreated(model.Category, document, cancellationToken);
-            }
-            catch
-            {
-                // Ignore failure to send notification about system event
-            }
-
+            var eventId = await _systemEventSender.Create(model, cancellationToken);
             return Ok(new EventResponse{
-                EventId = response.Id
+                EventId = eventId
             });
         }
 
@@ -96,39 +78,14 @@ namespace SystemEvents.Controllers
         [Route("event/start")]
         public async Task<ActionResult<EventResponse>> StartEvent([FromBody] EventRequestModel model, CancellationToken cancellationToken)
         {
-            if (!IsValid(model, out string reason))
+            if (!_systemEventSender.Validate(model, out string reason))
             {
                 return BadRequest($"Invalid request: {reason}");
             }
 
-            var response = await CreateSystemEvent(model, cancellationToken);
-            if (!response.IsValid)
-            {
-                _logger.LogDebug(response.DebugInformation);
-                _logger.LogError(response.OriginalException, "System event was not created");
-                return BadRequest("Unable to start a system event");
-            }
-
-            if (_categorySubscriptionNotifier == null)
-            {
-                return Ok(new EventResponse{
-                    EventId = response.Id
-                });
-            }
-
-            // Notify about this event
-            try
-            {
-                var document = await _esClient.GetAsync(response.Id, cancellationToken);
-                await _categorySubscriptionNotifier.OnEventStarted(model.Category, document, cancellationToken);
-            }
-            catch
-            {
-                // Ignore failure to send notification about system event
-            }
-
+            var eventId = await _systemEventSender.Start(model, cancellationToken);
             return Ok(new EventResponse{
-                EventId = response.Id
+                EventId = eventId
             });
         }
 
@@ -147,129 +104,8 @@ namespace SystemEvents.Controllers
                 return BadRequest($"{nameof(eventId)} can not be null or whitespace");
             }
 
-            var response = await _esClient.UpdateAsync(
-                        eventId,
-                        new SystemEventElasticsearchPartialDocument { Endtime = _timeStampFactory.GetTimestamp()},
-                        cancellationToken: cancellationToken, retryWithPreviousIndex: true);
-
-            if (!response.IsValid)
-            {
-                _logger.LogError(response.OriginalException, "System event was not updated");
-                return BadRequest("Unable to update system event");
-            }
-
-            if (_categorySubscriptionNotifier == null)
-            {
-                return Ok();
-            }
-
-            // Notify about this event
-            try
-            {
-                var document = await _esClient.GetAsync(response.Id, cancellationToken);
-                await _categorySubscriptionNotifier.OnEventFinished(document.Category, document, cancellationToken);
-            }
-            catch
-            {
-                // Ignore failure to send notification about system event
-            }
-
+            await _systemEventSender.End(eventId, cancellationToken);
             return Ok();
-        }
-
-        private Task<IIndexResponse> CreateSystemEvent(EventRequestModel model, CancellationToken cancellationToken)
-        {
-            if (model.Tags == null)
-            {
-                model.Tags = new List<string>();
-            }
-            
-            model.Tags.Add(model.Level.ToString());
-            model.Tags.Add(model.Level.ToString());
-
-            var remoteIpAddress = HttpContext.Connection.RemoteIpAddress;
-            var eventTimestamp = _timeStampFactory.GetTimestamp();
-
-            var systemEvent = new SystemEventElasticsearchDocument
-            {
-                Category = model.Category,
-                Level = model.Level.ToString(),
-                TargetKey = model.TargetKey,
-                Message = $"{model.Message} by {model.Sender}",
-                Tags = model.Tags,
-                Sender = model.Sender,
-                RemoteIpAddress = remoteIpAddress.ToString(),
-                Timestamp = eventTimestamp,
-                Endtime = eventTimestamp
-            };
-
-            return _esClient.IndexAsync(systemEvent, cancellationToken: cancellationToken);
-        }
-
-        private bool IsValid(EventRequestModel model, out string reason)
-        {
-            if (model == null)
-            {
-                reason = $"{nameof(model)} can not be null";
-                return false;
-            }
-            
-            if (string.IsNullOrWhiteSpace(model.Category))
-            {
-                reason = $"{nameof(model.Category)} can not be null or whitespace";
-                return false;
-            }
-
-            if (string.IsNullOrWhiteSpace(model.TargetKey))
-            {
-                reason = $"{nameof(model.TargetKey)} can not be null or whitespace";
-                return false;
-            }
-            
-            if (string.IsNullOrWhiteSpace(model.Message))
-            {
-                reason = $"{nameof(model.Message)} can not be null or whitespace";
-                return false;
-            }
-            
-            if (string.IsNullOrWhiteSpace(model.Sender))
-            {
-                reason = $"{nameof(model.Sender)} can not be null or whitespace";
-                return false;
-            }
-
-            model.Category = model.Category.Trim();
-            model.TargetKey = model.TargetKey.Trim();
-            model.Sender = model.Sender.Trim();
-
-            if (_advanceConfiguration?.Categories == null)
-            {
-                reason = null;
-                return true;
-            }
-
-            var allAllowed = _advanceConfiguration.Categories.Any(c => c.Name == "*");
-
-            // If Advance Configuration is enabled then only specified categories 
-            // from the configuration can be used
-            var category = _advanceConfiguration.Categories.FirstOrDefault(
-                                                  c => c.Name == model.Category);
-            if (!allAllowed && category == null)
-            {
-                reason = $"The provided category `{model.Category}` is not allowed. Check" + 
-                        "/category/all for a list of allowed categories or contact your system administrator.";
-                return false;
-            }
-
-            if (category?.Level != null && category.Level.Value != model.Level)
-            {
-                reason = $"Only events of level `{category.Level.Value}` are allowed for category `{model.Category}`. Check" + 
-                        "/category/all for a list of allowed categories or contact your system administrator.";
-                return false;
-            }
-            
-            reason = null;
-            return true;
         }
     }
 }