+++ system-events/src/Enums/SlackCommand.cs
@@ -1,0 +1,8 @@
+namespace SystemEvents.Enums
+{
+    public enum SlackCommand
+    {
+        Create,
+        End
+    }
+}