+++ system-events/src/Models/Slack/OpenSlackModalRequest.cs
@@ -1,0 +1,8 @@
+namespace SystemEvents.Models.Slack
+{
+    public class OpenSlackModalRequest
+    {
+        public string trigger_id { get; set; }
+        public string view { get; set; }
+    }
+}